% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_finna.R
\name{search_finna}
\alias{search_finna}
\title{Search Finna Index with Extended Options}
\usage{
search_finna(
  lookfor,
  type = "AllFields",
  fields = NULL,
  filters = NULL,
  facets = NULL,
  facetFilters = NULL,
  sort = "relevance,id asc",
  page = 1,
  limit = 100,
  lng = "fi",
  prettyPrint = FALSE
)
}
\arguments{
\item{lookfor}{Search terms. May include boolean operators.}

\item{type}{Search type (e.g., "AllFields", "Title", "Author", "Subject"). Defaults to "AllFields".}

\item{fields}{A vector of fields to return. Defaults to NULL, which returns all default fields.}

\item{filters}{A vector of filter queries. Defaults to NULL.}

\item{facets}{A vector of facets queries. Defaults to NULL.}

\item{facetFilters}{A vector of facetFilters queries. Defaults to NULL}

\item{sort}{Sort method. Defaults to "relevance,id asc".}

\item{page}{The page number to retrieve. Defaults to 1.}

\item{limit}{The number of records to return per page. Defaults to 20.}

\item{lng}{Language for returned translated strings. Defaults to "fi".}

\item{prettyPrint}{Logical; whether to pretty-print the response. Defaults to FALSE.}
}
\value{
A list containing the search results.
}
\description{
Search Finna Index with Extended Options
}
\examples{
search_results <- search_finna("sibelius")
print(search_results)
}
